refactor(ci-tools): narrow bootstrap-safe workflow-report exports; shrink bootstrap-closure baseline to weaver residual
Extract the dependency-free workflow-report wire contract out of the wide
@overeng/ci-tools barrel (src/mod.ts export *'d the runtime ./deploy-* encoders
alongside it) into a new src/workflow-report.ts that imports nothing runtime.
mod.ts re-exports it, so runtime consumers are unaffected. Repoint the genie
helpers (ci-workflow/reporting.ts, deploy-preview/shared.ts) to the narrow
module so they no longer drag effect into their bootstrap import closure.

Also narrow tsconfig.all.json.genie.ts from the wide @overeng/genie/node barrel
(which re-exports the typescript-importing closure walker) to the bootstrap-safe
node/tsconfig-from-packages.ts module.

These clear 74 of 79 violations (73 effect + 1 typescript); regenerate the
bootstrap-closure baseline to the 5 genie/weaver-registry/*.genie.ts residual,
which reach effect because the OTel semconv contracts are Effect-Schema by
design and render post-install. Whether to exempt or re-architect those is an
open decision.
