+++ genie/bootstrap-closure-baseline.json
@@ -1,84 +1,10 @@
 {
   "description": "Accepted (pre-existing) bootstrap-closure violations — repo-relative `.genie.ts` sources that transitively reach a runtime-only package. Baseline + ratchet: a NEW violation fails the gate; a fixed source should be removed here. Regenerate with `bootstrap-closure:check --update-baseline`.",
   "sources": [
-    ".github/workflows/auto-review.yml.genie.ts",
-    ".github/workflows/ci.yml.genie.ts",
-    "context/effect/socket/package.json.genie.ts",
-    "context/effect/socket/tsconfig.json.genie.ts",
-    "context/opentui/package.json.genie.ts",
-    "context/opentui/tsconfig.json.genie.ts",
     "genie/weaver-registry/attributes.yaml.genie.ts",
     "genie/weaver-registry/constants.rs.genie.ts",
     "genie/weaver-registry/constants.ts.genie.ts",
     "genie/weaver-registry/manifest.yaml.genie.ts",
-    "genie/weaver-registry/signals.yaml.genie.ts",
-    "package.json.genie.ts",
-    "packages/@overeng/agent-session-ingest/package.json.genie.ts",
-    "packages/@overeng/agent-session-ingest/tsconfig.json.genie.ts",
-    "packages/@overeng/ci-tools/package.json.genie.ts",
-    "packages/@overeng/ci-tools/tsconfig.json.genie.ts",
-    "packages/@overeng/content-address/package.json.genie.ts",
-    "packages/@overeng/content-address/tsconfig.json.genie.ts",
-    "packages/@overeng/effect-ai-claude-cli/package.json.genie.ts",
-    "packages/@overeng/effect-ai-claude-cli/tsconfig.json.genie.ts",
-    "packages/@overeng/effect-path/package.json.genie.ts",
-    "packages/@overeng/effect-path/tsconfig.json.genie.ts",
-    "packages/@overeng/effect-react/package.json.genie.ts",
-    "packages/@overeng/effect-react/tsconfig.json.genie.ts",
-    "packages/@overeng/effect-rpc-tanstack/examples/basic/package.json.genie.ts",
-    "packages/@overeng/effect-rpc-tanstack/examples/basic/tsconfig.json.genie.ts",
-    "packages/@overeng/effect-rpc-tanstack/package.json.genie.ts",
-    "packages/@overeng/effect-rpc-tanstack/tsconfig.json.genie.ts",
-    "packages/@overeng/effect-schema-form-aria/package.json.genie.ts",
-    "packages/@overeng/effect-schema-form-aria/tsconfig.json.genie.ts",
-    "packages/@overeng/effect-schema-form/package.json.genie.ts",
-    "packages/@overeng/effect-schema-form/tsconfig.json.genie.ts",
-    "packages/@overeng/genie/package.json.genie.ts",
-    "packages/@overeng/genie/tsconfig.json.genie.ts",
-    "packages/@overeng/kdl-effect/package.json.genie.ts",
-    "packages/@overeng/kdl-effect/tsconfig.json.genie.ts",
-    "packages/@overeng/kdl/package.json.genie.ts",
-    "packages/@overeng/kdl/tsconfig.json.genie.ts",
-    "packages/@overeng/megarepo/package.json.genie.ts",
-    "packages/@overeng/megarepo/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-cli/package.json.genie.ts",
-    "packages/@overeng/notion-cli/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-core/package.json.genie.ts",
-    "packages/@overeng/notion-core/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-datasource-sync/package.json.genie.ts",
-    "packages/@overeng/notion-datasource-sync/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-effect-client/package.json.genie.ts",
-    "packages/@overeng/notion-effect-client/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-effect-schema/package.json.genie.ts",
-    "packages/@overeng/notion-effect-schema/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-md/package.json.genie.ts",
-    "packages/@overeng/notion-md/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-property-write/package.json.genie.ts",
-    "packages/@overeng/notion-property-write/tsconfig.json.genie.ts",
-    "packages/@overeng/notion-react/package.json.genie.ts",
-    "packages/@overeng/notion-react/tsconfig.json.genie.ts",
-    "packages/@overeng/otel-contract/package.json.genie.ts",
-    "packages/@overeng/otel-contract/tsconfig.json.genie.ts",
-    "packages/@overeng/oxc-config/package.json.genie.ts",
-    "packages/@overeng/oxc-config/tsconfig.json.genie.ts",
-    "packages/@overeng/pty-effect/package.json.genie.ts",
-    "packages/@overeng/pty-effect/tsconfig.json.genie.ts",
-    "packages/@overeng/react-inspector/package.json.genie.ts",
-    "packages/@overeng/react-inspector/tsconfig.json.genie.ts",
-    "packages/@overeng/restate-effect/package.json.genie.ts",
-    "packages/@overeng/restate-effect/tsconfig.json.genie.ts",
-    "packages/@overeng/tui-core/package.json.genie.ts",
-    "packages/@overeng/tui-core/tsconfig.json.genie.ts",
-    "packages/@overeng/tui-react/package.json.genie.ts",
-    "packages/@overeng/tui-react/tsconfig.json.genie.ts",
-    "packages/@overeng/tui-stories/package.json.genie.ts",
-    "packages/@overeng/tui-stories/tsconfig.json.genie.ts",
-    "packages/@overeng/utils-dev/package.json.genie.ts",
-    "packages/@overeng/utils-dev/tsconfig.json.genie.ts",
-    "packages/@overeng/utils/package.json.genie.ts",
-    "packages/@overeng/utils/tsconfig.json.genie.ts",
-    "pnpm-install-contract.json.genie.ts",
-    "pnpm-workspace.yaml.genie.ts",
-    "tsconfig.all.json.genie.ts"
+    "genie/weaver-registry/signals.yaml.genie.ts"
   ]
 }
