+++ genie/deploy-preview/shared.ts
@@ -1,4 +1,4 @@
-import { workflowReportRecordLineMarker } from '../../packages/@overeng/ci-tools/src/mod.ts'
+import { workflowReportRecordLineMarker } from '../../packages/@overeng/ci-tools/src/workflow-report.ts'
 
 export const deployTargetEnvSuffix = (name: string) =>
   name
