+++ packages/@overeng/genie/docs/spec.md
@@ -201,6 +201,8 @@ Satisfies R06 (runtime independence of the transitive closure) and R30 (bootstra
 
 **Enforcement (baseline + ratchet).** Pre-existing violations are captured in a committed baseline of source paths so the gate is green today; only _new_ violations fail. The check warns on stale baseline entries (a baselined source that no longer violates) so the baseline ratchets down as violations are fixed. It runs as a devenv task (`bootstrap-closure:check`) wired into `check:all` (not `check:quick`) — an existing check-flow surface, not a new Genie CLI subcommand. The `checkBootstrapClosure` walker is exported from `@overeng/genie/node`, so downstream megarepo members can enforce the same contract over their own `.genie.ts` (whose closures do reach into effect-utils via `#mr/effect-utils/...`).
 
+The baseline has since been ratcheted from its initial 79 entries down to the **5 `genie/weaver-registry/*.genie.ts` residual**. The mechanically-fixable root causes were fixed at the source (a wide `@overeng/ci-tools` barrel that `export *`ed the runtime `./deploy-*` encoders alongside the bootstrap-safe workflow-report constants — narrowed by extracting the safe surface into `ci-tools/src/workflow-report.ts` and repointing genie helpers there; and `tsconfig.all.json.genie.ts` importing the wide `@overeng/genie/node` barrel — narrowed to the bootstrap-safe `node/tsconfig-from-packages.ts` module). The remaining 5 reach `effect` because the OTel semantic-convention contracts are Effect-Schema by design (`@overeng/otel-contract/src/registry.ts` imports `Schema` from `effect`); these weaver generators render under the normal `genie:run` (gated `after pnpm:install`), so their runtime-only closure is a post-install-generator property, not an accidental barrel leak. Whether to formally exempt post-install-Schema generators from the fresh-clone contract, or re-architect them onto a dep-free surface, is an open decision — the residual baseline holds the line until it is made.
+
 Relates to DQ1: this is the validate-only flavor of the preflight phase DQ1 anticipates; it detects unavailable closures but does not itself materialize members.
 
 ## Discovery and Validation
